Move LockSpec Models to models/
