+++ conda_lock/src_parser/environment_yaml.py
@@ -6,8 +6,8 @@ from typing import List, Optional, Sequence, Tuple
 
 import yaml
 
+from conda_lock.models.lock_spec import Dependency, LockSpecification
 from conda_lock.src_parser.conda_common import conda_spec_to_versioned_dep
-from conda_lock.src_parser.models import Dependency, LockSpecification
 from conda_lock.src_parser.selectors import filter_platform_selectors
 
 from .pyproject_toml import parse_python_requirement
